[HatoholArmPluginBase] Suppress a warning.
Without this patch, the following warning is shown.

HatoholArmPluginBase.cc: In member function 'mlpl::SmartTime HatoholArmPluginBase::getTimestampOfLastTrigger()':
HatoholArmPluginBase.cc:107:61: warning: narrowing conversion of 'body->HapiResTimestampOfLastTrigger::nanosec' from 'const uint32_t {aka const unsigned int}' to '__syscall_slong_t {aka long int}' inside { } is ill-formed in C++11 [-Wnarrowing]
  const timespec ts = {(time_t)body->timestamp, body->nanosec};
                                                             ^

+++ server/src/HatoholArmPluginBase.cc
@@ -104,7 +104,7 @@ SmartTime HatoholArmPluginBase::getTimestampOfLastTrigger(void)
 
 	const HapiResTimestampOfLastTrigger *body = 
 	  getResponseBody<HapiResTimestampOfLastTrigger>(m_ctx->responseBuf);
-	const timespec ts = {(time_t)body->timestamp, body->nanosec};
+	const timespec ts = {(time_t)body->timestamp, (long)body->nanosec};
 	return SmartTime(ts);
 }
